front/syn_metadata: move pure type-string helpers and dedupe first_top_level_generic_arg (Slice 2)
